version 1.4.2

+++ internal/app/model/shipmentbysap.go
@@ -9,7 +9,7 @@ func Main() *Shipmentbysaps {
 
 // Shipmentbysap ...
 type Shipmentbysap struct {
-	Material     int
+	Material     int64
 	Qty          int64
 	ShipmentDate time.Time
 	ShipmentTime time.Time
